projects/dsb/cup/pyproject.toml
[project]
name = "cup"
version = "0.1.0"
description = "PDF Processing Tools - Extract and convert PDF content to various formats"
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
  "click==8.2.1",
  "langchain==0.3.27",
  "langchain-openai==0.3.28",
  "numpy==2.3.2",
  "openai==1.98.0",
  "opencv-python-headless==4.12.0.88",
  "polars==1.32.0",
  "pdf2image==1.17.0",
  "pillow==11.3.0",
  "pypdf==5.9.0",
  "pydantic==2.11.7",
  "rich==14.1.0",
  "surya-ocr==0.14.7",
  "tabula-py[jpype]==2.10.0",
  "tabulate==0.9.0",
  "typer==0.16.0",
]

[project.scripts]
pdf-extract = "src.cli:app"
pdf-convert = "src.cli:app"

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "black==25.1.0",
]

[tool.uv]
override-dependencies = ["numpy==2.3.2", "pillow==11.3.0"]
